refactor: abstract AI service
